Eliminated some warnings

+++ src/simage_rgb.c
@@ -81,7 +81,7 @@ ConvertShort(unsigned short *array, long length)
   while (length--) {
     b1 = *ptr++;
     b2 = *ptr++;
-    *array++ = (b1 << 8) | (b2);
+    *array++ = (unsigned short) ((b1 << 8) | (b2));
   }
 }
